Started adding the instantiation for Tree
input manager takes a mouse click, single event, which is uses to get
reference position for setting the root of the RTT algorithm search.
Next step is to take this forward to check if the mouse position is
inside map, check it is not colliding with a '@' symbol, and draw the
root node.