Draw Nodes, Relative Position bug
Drawing root node is in place and setting new position for root
(eventually used for calculating path between two points) works. There
is a bug which is stopping this from working when screen view moves,
position does not seem to be truly relative. Also, needs to take
zoomLevel into account.