Began the texture version, started node-generation
Began the long haul of converting the map into a single texture drawn as
a sprite on screen. This is a long process to convert but is HUGELY
impacting on the FPS as it eliminates any bottleneck I was suffering
with. I have also began the RTT algorithm and started the node
generation, although right now it is purely adding nodes to the map in a
randomised order and not connecting them